+++ semana2/ejercicio1.txt
@@ -1,0 +1,40 @@
+1. Cree un pseudocódigo que le pida un precio de producto al usuario, calcule su descuento y muestre el precio final tomando en cuenta que:
+    1. Si el precio es menor a 100, el descuento es del 2%.
+    2. Si el precio es mayor o igual a 100, el descuento es del 10%.
+    3. *Ejemplos*:
+        1. 120 → 108
+        2. 40 → 39.2
+
+Inicio
+
+Definir precio
+
+Definir descuento
+
+Definir precio_final
+
+Mostrar “Ingrese precio de producto”
+
+Pedir precio
+
+Si (precio< 100) entonces
+
+Descuento=precio*0.02
+
+precio_final=Descuento-precio
+
+`Sino`
+
+Si(precio>=100) entonces
+
+Descuento=precio*0.10
+
+precio_final=Descuento-precio
+
+FinSi
+
+Mostrar” El precio de su producto sería”
+
+Mostrar precio_final
+
+Fin
