+++ semana2/ejercicio3.txt
@@ -1,0 +1,32 @@
+3. Cree un algoritmo que le pida un numero al usuario, y realice una suma de cada numero del 1 hasta ese número ingresado. Luego muestre el resultado de la suma.
+    1. 3 → 6 (1 + 2 + 3)
+    2. 5 → 15 (1 + 2 + 3 + 4 + 5)
+    3. 12 → 78 (1 + 2 + 3 + 4 + 5 + 6 + 7 + 8 + 9 + 10 + 11 + 12)
+
+Inicio
+
+Definir numero
+
+Definir contador
+
+Definir suma
+
+suma=1
+
+Mostrar “Ingrese número”
+
+Pedir numero
+
+Mientras que(suma<=numero) hacer:
+
+suma=suma+contador———>Aquí debe ser suma=suma+contador
+
+contador = contador+1——>Excelente aquí con el contador
+
+FinMientras
+
+Mostrar” El resultado sería:”
+
+Mostrar suma
+
+Fin
