+++ semana2/ejercicio4.txt
@@ -1,0 +1,41 @@
+1. Cree un algoritmo que le pida 2 números al usuario, los guarde en dos variables distintas (primero y segundo) y los ordene de menor a mayor en dichas variables.
+    1. Ejemplos:
+        1. A: 56, B: 32 → A: 32, B: 56
+        2. A: 24, B: 76 → A: 24, B: 76
+        3. A: 45, B: 12 → A: 12, B: 45
+
+Inicio
+
+Definir uno
+
+Definir dos
+
+Definir aux
+
+Mostrar” Ingrese el primer número: ”
+
+Pedir uno
+
+Mostrar “Ingrese el segundo número: ”
+
+Pedir dos
+
+Mientras que (uno < dos) hacer:
+
+Si (uno>dos) entonces:
+
+aux=uno
+
+uno=dos
+
+dos=aux
+
+Finsi
+
+Finmientras
+
+Mostrar uno
+
+Mostrar dos
+
+Fin
