+++ semana2/ejercicio5.txt
@@ -1,0 +1,102 @@
+1. Cree un algoritmo que le pida al usuario una velocidad en km/h y la convierta a m/s. Recuerda que 1 km == 1000m y 1 hora == 60 minutos * 60 segundos.
+    1. *Ejemplos*:
+        1. 73 → 20.27
+        2. 50 → 13.88
+        3. 120 → 33.33
+
+Inicio
+
+Definir km
+
+Definir metros
+
+Definir ms
+
+Definir total
+
+metros=1000
+
+segundos=3600
+
+Mostrar “Agregue los kilómetros”
+
+Pedir km
+
+ms= km*metros/segundos
+
+Mostrar “Los metros sobre segundos serian :”
+
+Mostrar ms
+
+Fin
+
+3.Cree un algoritmo que le pregunte al usuario por el sexo de 6 personas, ingresando 1 si es mujer o 2 si es hombre, y muestre al final el porcentaje de mujeres y hombres.
+
+1. *Ejemplos*:
+    1. 1, 1, 1, 2, 2, 2 → 50% mujeres y 50% hombres
+    2. 1, 1, 2, 2, 2, 2 → 33.3% mujeres y 66.6% hombres
+    3. 1, 1, 1, 1, 1, 2 → 84.4% mujeres y 16.6% hombres
+
+Inicio
+
+Definir mujer
+
+Definir hombre
+
+Definir persona
+
+Definir porcentual
+
+Definir total1
+
+Definir total2
+
+Mostrar “2 para hombre”
+
+Pedir hombre
+
+Mostrar “Al ingresar presiona 1”
+
+Pedir persona
+
+Mientras que(persona<=6) hacer:
+
+Mostrar ”Seleccione su sexo”
+
+Mostrar “1 para mujer”
+
+Pedir mujer
+
+Mostrar “2 para hombre”
+
+Pedir hombre
+
+Si (mujer==1) entonces
+
+mujer=mujer+1
+
+`Sino`
+
+Si(hombre==2)
+
+hombre=hombre+1
+
+Finsi
+
+FinMientras
+
+porcentual=mujeres+hombres
+
+total1= mujeres/porcental*100
+
+total2= hombres/porcental*100
+
+Mostrar ”El porcentaje de mujeres es: ”
+
+Mostrar total1
+
+Mostrar “El porcentaje de hombres es”
+
+Mostrar total2
+
+Fin
