+++ semana3/ejercicio1.py
@@ -1,0 +1,4 @@
+print("Hello World")
+print("Lyfter rules")
+print("Updating our minds!")
+print("Wat")
