+++ semana3/ejercicio2.py
@@ -1,8 +1,3 @@
-print("Hello World")
-print("Lyfter rules")
-print("Updating our minds!")
-print("Wat")
-
 print("El resultado de 1+1 es =",1+1),
 print("El resultado de 90-30 es =", 90-30),
 print("El resultado de 10*4 es =",10*4)
