+++ semana5/DictSemana5/ejercicio4.py
@@ -1,0 +1,74 @@
+#4 Dada una lista de ventas con la siguiente información:
+#date
+#customer_email
+#items
+#Y cada item teniendo la siguiente información:
+#name
+#upc
+#unit_price
+#Cree un diccionario que guarde el total de ventas de cada UPC.
+sales = [
+	{
+		'date': '27/02/23',
+		'customer_email': 'joe@gmail.com',
+		'items': [
+			{
+				'name': 'Lava Lamp',
+				'upc': 'ITEM-453',
+				'unit_price': 65.76,
+			},
+			{
+				'name': 'Iron',
+				'upc': 'ITEM-324',
+				'unit_price': 32.45,
+			},
+			{
+				'name': 'Basketball',
+				'upc': 'ITEM-432',
+				'unit_price': 12.54,
+			},
+		],
+	},
+	{
+		'date': '27/02/23',
+		'customer_email': 'david@gmail.com',
+		'items': [
+			{
+				'name': 'Lava Lamp',
+				'upc': 'ITEM-453',
+				'unit_price': 65.76,
+			},
+			{
+				'name': 'Key Holder',
+				'upc': 'ITEM-23',
+				'unit_price': 5.42,
+			},
+		],
+	},
+	{
+		'date': '26/02/23',
+		'customer_email': 'amanda@gmail.com',
+		'items': [
+			{
+				'name': 'Key Holder',
+				'upc': 'ITEM-23',
+				'unit_price': 3.42,
+			},
+			{
+				'name': 'Basketball',
+				'upc': 'ITEM-432',
+				'unit_price': 17.54,
+			},
+		],
+	},
+]
+
+Total={}
+for sell in sales:
+    for item in sell['items']:
+        upc=item['upc']
+        cost=item['unit_price']
+        Total[upc]=Total.get(upc,0)+cost
+
+for UP,result in Total.items():
+    print(f'UPC:{UP},Total en ventas:${result}')
