+++ semana5/ejercicio2d.py
@@ -1,0 +1,7 @@
+#ejercicio2
+#Cree un programa que itere e imprima un string letra por letra de derecha a izquierda.
+
+string = 'Pizza con piña'
+
+for char in reversed(string):
+    print(char)
