+++ semana6/pract1.py
@@ -1,0 +1,20 @@
+#1.Cree dos funciones que impriman dos cosas distintas, y haga que la primera llame la segunda.
+def first_function():
+    incredible=2+2
+    print(f'Programing rules {incredible}!')
+    second_function()
+
+def second_function():
+    print('This is amazing')
+
+first_function()
+
+
+
+
+
+
+
+    
+
+
